Change date format

+++ packages/date-logic/index.js
@@ -1,6 +1,6 @@
 module.exports = function getDate(
     locale = 'en-US',
-    options = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' }
+    options = { weekday: 'short', year: 'numeric', month: 'short', day: 'numeric' }
 ) {
     return new Date().toLocaleDateString(locale, options);
 }
